Further watermark improvements:
- refactored how watermarking is done to reduce the number of code branches
- fixed the text scaling
